restructuring step 2
